separate specs and e2e tests
