+++ test/specs/models.js
@@ -1,6 +1,6 @@
 var chai = require('chai');
 var should = chai.should();
-var User = require('../models/User');
+var User = require('../../models/User');
 
 describe('User Model', function() {
   it('should create a new user', function(done) {
